requirements.txt
python-docx>=1.1.0
flask>=3.0.0
lxml>=5.0.0
gunicorn>=23.0.0
